// Codice/BMT - Business Management Tool/Richiesta 2/src/spas/app/redux-store/slices/ui/ui.interfaces.ts
export enum DialogTypes {
  CREATE_CUSTOMER = "CREATE_CUSTOMER",
  EDIT_CUSTOMER = "EDIT_CUSTOMER",
  CREATE_PROJECT = "CREATE_PROJECT",
  EDIT_PROJECT = "EDIT_PROJECT",
  CREATE_JOB = "CREATE_JOB",
  CREATE_ACTIVITY = "CREATE_ACTIVITY",
  EDIT_JOB = "EDIT_JOB",
  CREATE_WORKER = "CREATE_WORKER",
  EDIT_WORKER = "EDIT_WORKER",
  CREATE_EDIT_WORKER_CREDENTIALS = "CREATE_EDIT_WORKER_CREDENTIALS",
  CREATE_WORKING_DAY = "CREATE_WORKING_DAY",
  EDIT_WORKING_DAY = "EDIT_WORKING_DAY",
  ASSOCIATE_JOBS = "ASSOCIATE_JOBS",
  CHECK_IMPORTED_ROWS = "CHECK_IMPORTED_ROWS",
  CREATE_COLLABORATION_PAYMENT = "CREATE_COLLABORATION_PAYMENT",
  EDIT_COLLABORATION_PAYMENT = "EDIT_COLLABORATION_PAYMENT",
  CREATE_EMPLOYEE_PAYMENT = "CREATE_EMPLOYEE_PAYMENT",
  EDIT_EMPLOYEE_PAYMENT = "EDIT_EMPLOYEE_PAYMENT",
  CREATE_JOB_PAYMENT = "CREATE_JOB_PAYMENT",
  EDIT_JOB_PAYMENT = "EDIT_JOB_PAYMENT",
  CREATE_TAX = "CREATE_TAX",
  EDIT_TAX = "EDIT_TAX",
  CREATE_SERVICE_EXPENSE = "CREATE_SERVICE_EXPENSE",
  EDIT_SERVICE_EXPENSE = "EDIT_SERVICE_EXPENSE",
  CREATE_EXTRA_INCOME = "CREATE_EXTRA_INCOME",
  EDIT_EXTRA_INCOME = "EDIT_EXTRA_INCOME",
  TRANSACTIONS_DETAIL = "TRANSACTIONS_DETAIL",
}

export interface UiState {
  isDialogOpen: {
    [key in DialogTypes]: boolean;
  };
}